Plot cell state graph

Plot cell state graph

Arguments

cell_colors
character vector indicating the metadata to use to color the vertices. Names can be either one of the column names of the phenotypic metadata structure (see colnames(pData(antler$expressionSet))), names of a cluster entries, or gene names. The element of the character vector can be named themselves, in that case the element name will serve as a title for phenotypic metadata or cluster entries. If gene names are specified, their vector names must indicate the type of read counts that will be used to render the expression levels: "Raw", "Normalized" or "Smoothed". If no gene name's name is specified raw level will be rendered.

labels
character vector. Labels to draw on top of the cells. Default to NA.

vertex.size
numeric vector indicating the size of the vertices representing the cells.

vertex.shape
character vector indicating the shape of the vertices. Can be either "circle" (default), "square" or "triangle".

vertex.frame.color
character vector indicating the color of the frame of each vertex. Default to NA.

edge.width
numeric value indicating the width of graph edges. Default to a fifth of the vertex size.

edge.color
character string indicating the color of the edges. Default to "grey".

width
numeric value indicating the width (in inches) of each individual subplots. Default to 3.

height
numeric value indicating the height (in inches) of each individual subplots. Default to 3.

title
logical value indicating whether to show each subplots' title. Default to TRUE.

save_pdf
logical value indicating whether to save plot as a pdf file. Default to TRUE.

shuffle
logical value indicating whether to randomize the z-order of each vertex. Default to FALSE.

vertex.plot.order
integer vector. The z-order used to render the vertices. Default to NA.

gene_palette
character string. The color palette to use if genes are specified as cell colors.

filename
character string. The filename of the output file, if any.

...
Extra arguments passed to plot.igraph.