Identify gene modules.

This method groups genes demonstrating concerted patterns of expression. TODO details from methods

Arguments

name
character string. The name of the list of gene modules. Multiple lists of gene modules under different names can be stored.

corr_matrix
numeric matrix. Specifies an optional correlation matrix. If not provided a Spearman correlation matrix will be calculated by the function.

corr_t
numeric value. Correlation cutoff value or correlation quantile cutoff value used to select genes. Either corr_t or corr_num_genes must be specified.

corr_num_genes
integer value. Approximate number of genes to select in the correlation filtering step. Either corr_t or corr_num_genes must be specified.

corr_quantile
logical. If FALSE (default), corr_t specifies the correlation threshold. If TRUE, corr_t is the percentile correlation threshold over all corr_matrix correlation values.

corr_min
integer value. The minimum number of high gene- gene correlation values a gene needs to have to be kept. Default to 3.

mod_min_cell
an integer value indicating the minimum number of expressed cells a gene module needs to have to be selected. Default to 0.

mod_max_cell
an integer value indicating the maximum number of expressed cells a gene module needs to have to be selected. Default to Inf.

mod_consistency_thres
numeric value. If in a gene module the ratio of expressed genes in the "positive" cells (see Details) is less than this value, the gene module is excluded. Default to 0.4.

min_cell_level
numeric value. If in a gene module the average expression level of the "positive" cells is less than this value, the gene module is excluded. Default to 0.

mod_skewness_thres
numeric value. In each gene module the skewness of the cells' average expression level is calculated. If the skewness is lower than this value, the gene module is excluded. Default to -Inf.

ordering_correlation_thres
numeric value. if the Spearman correlation between the cells' average expression level of a gene module and the phenotypic metadata specified by the ordering parameter (usually the cells' timepoint) is lower than this value. The gene module is excluded. If equal to NA (Default), this filtering step is not applied.

ordering
a character string indicating the phenotypic metadata to use for the ordering correlation filtering step.

num_final_gms
integer value. Constraints the resulting number of gene modules after applying the filtering steps. If NULL (Default), a heuristic method is used to find the optimal number of gene modules.

num_max_final_gms
integer value. Caps the maximum number of resulting gene modules. If NULL (Default), no capping is applied.

num_min_final_gms
integer value. Caps the minimum number of resulting gene modules. If NULL (Default), no capping is applied.

num_initial_gms
integer value. Constraints the initial number of clustered gene modules, before applying the filtering steps. If NULL (Default), a heuristic method is used to find the optimal number of gene modules.

clustering_method
character string. The agglomeration method to be used by the hierarchical clustering algorithm (hclust). Default to "ward.D2".

display
logical. Whether to display intermediate processing plots. Default to FALSE.

num_cores.
an integer value indicating the number of cores to use.

verbose
a logical value indicating whether live processing information should be printed. Default to TRUE.

verbose_final
a logical value indicating whether processing information should be printed just before the function returns. Default to FALSE.

process_plots
a logical value indicating whether to render the heuristic trade-off determining the number of gene modules for each iteration. Default to FALSE.