Heatmap of the transcriptomic profile.

Once the gene modules and optionally the cell clusters have been calculated, this function produces a heatmap showing the gene expression levels.

Arguments

gene_module_name.
character string. The name of the gene module list shown along the rows.

cell_clusters_name
character string. The name of the cell clusters shown along the columns. If 'none' (default), the columns will be ordered as in the current dataset.

cell_side_colors
character vector indicating the metadata to show in the column annotation bar. Names can be either one of the column names of the phenotypic metadata structure (see pData(antler$expressionSet)), names of a cluster entries, or gene names. Default to 'cells_samples'.

cell_side_colors_hide_legend
character vector. The name of the phenotypic metadata to hide in the legend. Default to "none".

gene_side_colors
character vector. The name of the feature metadata to show in the row annotation bar. Default to "none".

data_status
character string. Specifies whether the gene expression levels displayed are raw ("Raw"), normalized ("Normalized") or have been imputed ("Smoothed"). Default to "Normalized".

gene_level_func
character string. Specifies whether the gene expression levels are rendered as is ('none'), log-transformed ('log') or z-scored log-transformed ('logscaled', default).

heatmap_palette_centered
a character vector storing a color map to use if the gene_level_func is set to 'none' or 'log'.

heatmap_palette
a character vector storing a color map to use if the gene_level_func is set to logscaled'.

color_limit_value
numeric value. The upper limit of the color map if gene_level_func is set to 'none' or 'log' (default NULL for no set limit), the absolute value of the upper and lower limits of the color map if gene_level_func is set to 'logscaled' (default to 2).

gene_labels_filter
character vector. Either a list of genes whose intersection with the gene modules will be displayed as paragraph, or 'all' for rendering each individual gene name on each row, or 'all_gms' for rendering each individual gene name plus the gene module id on each row, or 'naked' for no row labels at all.

num_gene_per_line
if gene_labels_filter is a list of genes, the maximum number of gene names to render on a single paragraph row. Default to 8.

fontsize
integer value indicating the plot font size. Default to 10.

subsampling_ratio
numeric value. The ratio of cells to render, see parameter subsampling_min_cells for an exception. Default to 1.

subsampling_min_cells
integer value. If a cell cluster object is specified with cell_clusters_name, the minimum of cells to render per cell cluster. Default to 20.

file_settings
a list of lists indicating the type and dimensions of the output plots. For each plot type, a list containing three arguments must be specified: 'type' (either 'pdf' or 'png'), 'width' and 'height', and a fourth one, 'res', can be set to define the resolution of png files. The unit of the pdf device dimension is inches and pixels for png device. The png resolution is 200 ppi (default) so the appearance of a w10xh15 pdf file is the same as a w2000xh3000 png file Default to list(list(type='pdf', width=15, height=15))

suffix
a character string that will be appended to the file name. Default to NULL.

...
Extra arguments passed to pheatmap.